# platform/katalog-manager-api/openapi.yaml
openapi: 3.0.3
info:
  title: Stube katalog-manager-api
  version: "1.0.0"
  description: >
    Neutral management API and first-run backend for Stube. Owns the catalog
    WRITE path (item edits, deletes, and the import scan that registers files
    the operator already owns on disk) plus service configuration. Processing
    work is dispatched asynchronously by emitting Kafka task events. This
    service is content-neutral: it never acquires, downloads, or fetches
    content. How files arrive on disk is out of scope.
servers:
  - url: /api/manage
    description: Route prefix the edge proxy forwards to this service.
...
paths:
...
  /config:
    get:
      tags: [config]
      summary: Current non-secret configuration
      responses:
        "200":
          description: Current configuration (no secrets).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "401":
          $ref: "#/components/responses/Unauthorized"
    put:
      tags: [config]
      summary: Update non-secret configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ConfigUpdate"
      responses:
        "200":
          description: Updated configuration.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Config"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Not yet configured; run first-run setup.
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        OIDC access token. The issuer and audience are configured from the
        operator's environment / discovery document.
...
  responses:
    Unauthorized:
      description: Missing or invalid bearer token.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    BadRequest:
      description: Invalid request body or parameters.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
  schemas:
...
    Config:
      type: object
      properties:
        configured:
          type: boolean
        displayName:
          type: string
        oidcIssuer:
          type: string
        oidcClientId:
          type: string
        libraryPath:
          type: string
    ConfigUpdate:
      type: object
      required: [displayName, oidcIssuer, oidcClientId, libraryPath]
      properties:
        displayName:
          type: string
        oidcIssuer:
          type: string
        oidcClientId:
          type: string
        libraryPath:
          type: string
...
    Error:
      type: object
      properties:
        error:
          type: string
security:
  - bearerAuth: []